some code refactoring and updated directory structure
