+++ internal/application/pong/Pong.go
@@ -1,0 +1,5 @@
+package application
+
+func Pong() string {
+	return "pong"
+}
