+++ internal/application/pong/Pong_test.go
@@ -1,0 +1,14 @@
+package application
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPong(t *testing.T) {
+
+	resp := Pong()
+
+	assert.Equal(t, resp, "pong")
+}
