--- internal/controller/Ping.go
@@ -1,9 +1,0 @@
-package controller
-
-import (
-	"net/http"
-)
-
-func Ping(w http.ResponseWriter, r *http.Request) {
-	w.Write([]byte("pong"))
-}
